simplifying adding event code
